Adding Incoming Integration Tests (#31)
Since 1r1's were introduced, `rules` was deprecated. Tests for incoming was moved to a new folder, `functions`.
I've added the GH issue to the SNS message once it's created (so it can be passed to the Slack message). I also removed some hard-coding in the function to rely on the SNS message attributes. This will further be revised.
I also checked to ensure the length of the record is 1, since we don't account for that yet. If SNS ever does send back more than one record, we can revise the code.